Setup the project
* Set package name -> org.kimp.mustep
* Update dependencies and Gradle Plugin
* Remove useless code(library-{kotlin,android})
* Target and compile android sdk is Tiramisu now
