+++ gradle/libs.versions.toml
@@ -1,20 +1,31 @@
 [versions]
-agp = "7.2.1"
+# AndroidX
+appcompat = "1.6.0-alpha04"
+constraint_layout = "2.2.0-alpha02"
+
+# Kotlin
+kotlin = "1.7.0-RC2"
+core_ktx = "1.9.0-alpha04"
+
+# Build
+agp = "7.4.0-alpha03"
+ktlint = "0.45.2"
+
+# Test
 androidx_test = "1.4.0"
 androidx_test_ext = "1.1.3"
-appcompat = "1.4.2"
-compile_sdk_version = "31"
-constraint_layout = "2.1.3"
-core_ktx = "1.7.0"
-detekt = "1.20.0"
-dokka = "1.6.21"
 espresso_core = "3.4.0"
 junit = "4.13.2"
-kotlin = "1.6.21"
-ktlint = "0.45.2"
+
+# Android SDK
+min_sdk_version = "26"
+compile_sdk_version = "android-Tiramisu"
+target_sdk_version = "Tiramisu"
+
+# Plugins
+dokka = "1.6.21"
+detekt = "1.21.0-RC1"
 ktlint_gradle = "10.2.1"
-min_sdk_version = "21"
-target_sdk_version = "31"
 benmanesversion = "0.42.0"
 
 [libraries]
