--- library-android/src/main/java/com/ncorti/kotlin/template/library/android/NotificationUtil.kt
@@ -1,41 +1,0 @@
-package com.ncorti.kotlin.template.library.android
-
-import android.app.Notification
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.os.Build
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
-
-private const val CHANNEL_ID = "default"
-private const val NOTIFICATION_ID = 42
-
-class NotificationUtil(context: Context) {
-
-    init {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val name = context.getString(R.string.channel_name)
-            val descriptionText = context.getString(R.string.channel_description)
-            val importance = NotificationManager.IMPORTANCE_DEFAULT
-            val channel = NotificationChannel(CHANNEL_ID, name, importance).apply {
-                description = descriptionText
-            }
-            val notificationManager: NotificationManager =
-                context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(channel)
-        }
-    }
-
-    fun showNotification(context: Context, title: String, message: String): Notification {
-        val builder = NotificationCompat.Builder(context, CHANNEL_ID)
-            .setSmallIcon(R.drawable.library_notification_icon)
-            .setContentTitle(title)
-            .setContentText(message)
-            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-
-        return builder.build().apply {
-            NotificationManagerCompat.from(context).notify(NOTIFICATION_ID, this)
-        }
-    }
-}
