--- library-android/src/main/res/values/strings.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="channel_name">Default</string>
-    <string name="channel_description">Default notifications for this app</string>
-</resources>
