--- library-kotlin/src/main/java/com/ncorti/kotlin/template/library/FactorialCalculator.kt
@@ -1,11 +1,0 @@
-package com.ncorti.kotlin.template.library
-
-object FactorialCalculator {
-
-    tailrec fun computeFactorial(input: Int, temp: Long = 1L): Long =
-        when {
-            input < 0 -> error("Factorial is not defined for negative numbers")
-            input == 0 -> temp
-            else -> computeFactorial(input - 1, temp * input)
-        }
-}
